tests(but): migrate rub tests to other commands
