+++ crates/but/tests/but/command/rub.rs
@@ -2,21 +2,11 @@ use snapbox::str;
 
 use crate::{
     command::util::{
-        branch_commit_cli_id_for_file, branch_commit_cli_ids,
-        commit_file_with_worktree_changes_as_two_hunks, commit_two_files_as_two_hunks_each,
+        branch_commit_cli_ids, commit_two_files_as_two_hunks_each,
         status_json_with_files as status_json,
     },
-    utils::{CommandExt, Sandbox},
+    utils::Sandbox,
 };
-
-fn assigned_uncommitted_file_env() -> anyhow::Result<Sandbox> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-
-    env.setup_metadata(&["A", "B"]);
-    env.file("a.txt", "arbitrary text\n");
-    env.but("rub zz:a.txt A").assert().success();
-    Ok(env)
-}
 
 fn stack_assigned_contains_file(
     status: &serde_json::Value,
@@ -46,17 +36,6 @@ fn uncommitted_contains_file(status: &serde_json::Value, file_path: &str) -> boo
         .any(|change| change["filePath"].as_str().unwrap() == file_path)
 }
 
-fn uncommitted_cli_id_for_file(status: &serde_json::Value, file_path: &str) -> Option<String> {
-    status["uncommittedChanges"]
-        .as_array()
-        .unwrap()
-        .iter()
-        .find_map(|change| {
-            (change["filePath"].as_str().unwrap() == file_path)
-                .then(|| change["cliId"].as_str().unwrap().to_string())
-        })
-}
-
 fn branch_commits_contain_file(
     status: &serde_json::Value,
     branch_name: &str,
@@ -71,1175 +50,6 @@ fn branch_commits_contain_file(
         .flat_map(|branch| branch["commits"].as_array().unwrap().iter())
         .flat_map(|commit| commit["changes"].as_array().unwrap().iter())
         .any(|change| change["filePath"].as_str().unwrap() == file_path)
-}
-
-fn committed_file_cli_id_for_file(
-    status: &serde_json::Value,
-    branch_name: &str,
-    file_path: &str,
-) -> Option<String> {
-    status["stacks"]
-        .as_array()
-        .unwrap()
-        .iter()
-        .flat_map(|stack| stack["branches"].as_array().unwrap().iter())
-        .find(|branch| branch["name"].as_str().unwrap() == branch_name)
-        .and_then(|branch| {
-            branch["commits"]
-                .as_array()
-                .unwrap()
-                .iter()
-                .find_map(|commit| {
-                    commit["changes"]
-                        .as_array()
-                        .unwrap()
-                        .iter()
-                        .find_map(|change| {
-                            (change["filePath"].as_str().unwrap() == file_path)
-                                .then(|| change["cliId"].as_str().unwrap().to_string())
-                        })
-                })
-        })
-}
-
-#[test]
-fn assign_uncommitted_file() -> anyhow::Result<()> {
-    let env = assigned_uncommitted_file_env()?;
-    env.but("diff A@{stack}:a.txt")
-        .assert()
-        .success()
-        .stderr_eq(snapbox::str![])
-        .stdout_eq(snapbox::str![[r#"
-─────────╮
-k:c a.txt│
-─────────╯
-     1│+arbitrary text
-
-"#]]);
-    Ok(())
-}
-
-#[test]
-fn uncommitted_file_to_uncommitted_area() -> anyhow::Result<()> {
-    let env = assigned_uncommitted_file_env()?;
-    env.but("rub A@{stack}:a.txt zz")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-Unstaged the only hunk in a.txt in a stack
-
-"#]])
-        .stderr_eq(str![""]);
-
-    env.but("diff zz:a.txt")
-        .assert()
-        .success()
-        .stderr_eq(snapbox::str![])
-        .stdout_eq(snapbox::str![[r#"
-─────────╮
-n:c a.txt│
-─────────╯
-     1│+arbitrary text
-
-"#]]);
-
-    Ok(())
-}
-
-#[test]
-fn uncommitted_file_to_branch() {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-
-    env.setup_metadata(&["A", "B"]);
-    commit_file_with_worktree_changes_as_two_hunks(&env, "A", "a.txt");
-
-    env.but("rub a.txt A")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-Staged all hunks in a.txt in the uncommitted area → [A].
-
-"#]])
-        .stderr_eq(str![""]);
-}
-
-#[test]
-fn uncommitted_file_by_path_prefix_to_branch() {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-
-    env.setup_metadata(&["A", "B"]);
-    commit_file_with_worktree_changes_as_two_hunks(&env, "A", "path/a.txt");
-
-    env.but("rub path/ A")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-Staged hunk(s) → [A].
-
-"#]])
-        .stderr_eq(str![""]);
-}
-
-#[test]
-fn committed_file_to_uncommitted_area() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-
-    env.setup_metadata(&["A", "B"]);
-    commit_two_files_as_two_hunks_each(&env, "A", "a.txt", "b.txt", "first commit");
-    commit_two_files_as_two_hunks_each(&env, "A", "a.txt", "b.txt", "second commit");
-
-    env.but("--format json status -f")
-        .allow_json()
-        .assert()
-        .success()
-        // .stderr_eq(snapbox::str![""])
-        .stdout_eq(snapbox::str![[r#"
-...
-{
-  "uncommittedChanges": [],
-  "stacks": [
-    {
-      "cliId": "i0",
-      "assignedChanges": [],
-      "branches": [
-        {
-          "cliId": "g0",
-          "name": "A",
-          "commits": [
-            {
-...
-              "changes": [
-                {
-                  "cliId": "1#0:n",
-                  "filePath": "a.txt",
-                  "changeType": "modified"
-                },
-                {
-                  "cliId": "1#0:p",
-                  "filePath": "b.txt",
-                  "changeType": "modified"
-                }
-              ]
-            },
-            {
-...
-              "changes": [
-                {
-                  "cliId": "1#1:n",
-                  "filePath": "a.txt",
-                  "changeType": "added"
-                },
-                {
-                  "cliId": "1#1:p",
-                  "filePath": "b.txt",
-                  "changeType": "added"
-                }
-              ]
-            },
-            {
-...
-              "changes": [
-                {
-                  "cliId": "t:t",
-                  "filePath": "A",
-                  "changeType": "added"
-                }
-              ]
-            }
-...
-
-"#]]);
-
-    env.but("rub 1#0:p zz")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-Uncommitted changes
-
-"#]])
-        .stderr_eq(str![""]);
-
-    // Verify that `status` reflects the move.
-    env.but("--format json status -f")
-        .allow_json()
-        .assert()
-        .success()
-        .stderr_eq(snapbox::str![""])
-        .stdout_eq(snapbox::str![[r#"
-{
-  "uncommittedChanges": [
-    {
-      "cliId": "p",
-      "filePath": "b.txt",
-      "changeType": "modified"
-    }
-  ],
-  "stacks": [
-    {
-      "cliId": "j0",
-      "assignedChanges": [],
-      "branches": [
-        {
-          "cliId": "g0",
-          "name": "A",
-          "commits": [
-            {
-...
-              "changes": [
-                {
-                  "cliId": "1#0:n",
-                  "filePath": "a.txt",
-                  "changeType": "modified"
-                }
-              ]
-            },
-            {
-...
-              "changes": [
-                {
-                  "cliId": "1#1:n",
-                  "filePath": "a.txt",
-                  "changeType": "added"
-                },
-                {
-                  "cliId": "1#1:p",
-                  "filePath": "b.txt",
-                  "changeType": "added"
-                }
-              ]
-            },
-            {
-...
-              "changes": [
-                {
-                  "cliId": "t:t",
-                  "filePath": "A",
-                  "changeType": "added"
-                }
-...
-    },
-    {
-      "cliId": "k0",
-      "assignedChanges": [],
-      "branches": [
-        {
-          "cliId": "h0",
-          "name": "B",
-          "commits": [
-            {
-...
-              "changes": [
-                {
-                  "cliId": "l:p",
-                  "filePath": "B",
-                  "changeType": "added"
-                }
-              ]
-            }
-...
-
-"#]]);
-
-    Ok(())
-}
-
-#[test]
-fn shorthand_uncommitted_hunk_to_uncommitted_area() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-
-    env.setup_metadata(&["A", "B"]);
-    commit_file_with_worktree_changes_as_two_hunks(&env, "A", "a.txt");
-
-    // Assign the change to A.
-    env.but("rub a.txt A").assert().success();
-
-    // Verify that the first hunk is j0, and move it to uncommitted.
-    env.but("diff A@{stack}:a.txt")
-        .assert()
-        .success()
-        .stderr_eq(snapbox::str![])
-        .stdout_eq(snapbox::str![[r#"
-─────────╮
-k:2 a.txt│
-─────────╯
-   1  │-first
-     1│+firsta
-   2 2│ line
-   3 3│ line
-   4 4│ line
-─────────╮
-k:e a.txt│
-─────────╯
-    6  6│ line
-    7  7│ line
-    8  8│ line
-    9   │-last
-       9│+lasta
-
-"#]]);
-    env.but("rub km:2 zz")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-Unstaged a hunk in a.txt in a stack
-
-"#]])
-        .stderr_eq(str![""]);
-
-    // Verify that only one hunk moved back to uncommitted ("a.txt" appears both in the
-    // uncommitted area and in a stack).
-    env.but("--format json status -f")
-        .allow_json()
-        .assert()
-        .success()
-        .stderr_eq(snapbox::str![])
-        .stdout_eq(snapbox::str![[r#"
-{
-  "uncommittedChanges": [
-    {
-      "cliId": "n",
-      "filePath": "a.txt",
-      "changeType": "modified"
-    }
-  ],
-  "stacks": [
-    {
-      "cliId": "k0",
-      "assignedChanges": [
-        {
-          "cliId": "k",
-          "filePath": "a.txt",
-          "changeType": "modified"
-        }
-      ],
-      "branches": [
-        {
-          "cliId": "g0",
-          "name": "A",
-...
-
-"#]]);
-
-    Ok(())
-}
-
-#[test]
-fn uncommitted_hunk_to_branch() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-
-    // Must set metadata to match the scenario
-    env.setup_metadata(&["A", "B"]);
-
-    commit_file_with_worktree_changes_as_two_hunks(&env, "A", "a.txt");
-
-    // Verify that the first hunk is j0, and move it to uncommitted.
-    env.but("diff a.txt")
-        .assert()
-        .success()
-        .stderr_eq(snapbox::str![])
-        .stdout_eq(snapbox::str![[r#"
-─────────╮
-n:2 a.txt│
-─────────╯
-   1  │-first
-     1│+firsta
-   2 2│ line
-   3 3│ line
-   4 4│ line
-─────────╮
-n:e a.txt│
-─────────╯
-    6  6│ line
-    7  7│ line
-    8  8│ line
-    9   │-last
-       9│+lasta
-
-"#]]);
-    env.but("rub nk:2 A")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-Staged a hunk in a.txt in the uncommitted area → [A].
-
-"#]])
-        .stderr_eq(str![""]);
-
-    // Verify that only one hunk was assigned ("a.txt" appears both in the
-    // uncommitted area and in a stack).
-    env.but("--format json status -f")
-        .allow_json()
-        .assert()
-        .success()
-        .stderr_eq(snapbox::str![])
-        .stdout_eq(snapbox::str![[r#"
-{
-  "uncommittedChanges": [
-    {
-      "cliId": "n",
-      "filePath": "a.txt",
-      "changeType": "modified"
-    }
-  ],
-  "stacks": [
-    {
-      "cliId": "k0",
-      "assignedChanges": [
-        {
-          "cliId": "k",
-          "filePath": "a.txt",
-          "changeType": "modified"
-        }
-      ],
-      "branches": [
-        {
-          "cliId": "g0",
-          "name": "A",
-...
-
-"#]]);
-
-    Ok(())
-}
-
-#[test]
-fn uncommitted_hunk_to_commit() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-
-    // Must set metadata to match the scenario
-    env.setup_metadata(&["A", "B"]);
-
-    commit_file_with_worktree_changes_as_two_hunks(&env, "A", "a.txt");
-
-    let target_cli_id = branch_commit_cli_ids(&status_json(&env)?, "A")[0].clone();
-    // The amended commit is identified by its change ID, from a freshly built
-    // map that knows the post-amend workspace.
-    env.but(format!("rub zz:a.txt:#0 {target_cli_id}"))
-        .assert()
-        .success()
-        .stdout_eq(str![[r#"
-Amended [..] → 1
-
-"#]])
-        .stderr_eq(str![""]);
-
-    // Verify that only one hunk was assigned ("a.txt" still appears in the
-    // uncommitted area because there is one hunk still unassigned).
-    env.but("--format json status -f")
-        .allow_json()
-        .assert()
-        .success()
-        .stderr_eq(snapbox::str![])
-        .stdout_eq(snapbox::str![[r#"
-{
-  "uncommittedChanges": [
-    {
-      "cliId": "n",
-      "filePath": "a.txt",
-      "changeType": "modified"
-    }
-  ],
-...
-
-"#]]);
-
-    // Verify that the commit indeed received the hunk.
-    snapbox::assert_data_eq!(
-        env.open_repo()
-            .rev_parse_single("A:a.txt")?
-            .object()?
-            .try_into_blob()?
-            .take_data(),
-        str![[r#"
-firsta
-line
-line
-line
-line
-line
-line
-line
-last
-
-"#]],
-    );
-
-    Ok(())
-}
-
-// Regression: filenames with dashes should not be misinterpreted as ranges.
-// Before the fix, "my-file.txt" would be split on '-' and treated as a range
-// from "my" to "file.txt", which would fail.
-#[test]
-fn filename_with_dash_not_treated_as_range() {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-
-    env.setup_metadata(&["A", "B"]);
-    env.file("my-file.txt", "arbitrary text\n");
-
-    // Staging by filename should work — the dash should NOT be interpreted as a range separator
-    env.but("stage my-file.txt A")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-Staged the only hunk in my-file.txt in the uncommitted area → [A].
-
-"#]])
-        .stderr_eq(str![""]);
-}
-
-// Tests for convenience commands
-
-#[test]
-fn uncommit_command_on_commit() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-
-    env.setup_metadata(&["A", "B"]);
-    commit_two_files_as_two_hunks_each(&env, "A", "a.txt", "b.txt", "first commit");
-
-    // Get the commit ID from status
-    let status_output = env.but("--format json status").allow_json().output()?;
-    let status_json: serde_json::Value = serde_json::from_slice(&status_output.stdout)?;
-    let commit_cli_id = status_json["stacks"][0]["branches"][0]["commits"][0]["cliId"]
-        .as_str()
-        .unwrap();
-
-    // Test uncommit command
-    env.but(format!("uncommit {commit_cli_id}"))
-        .assert()
-        .success();
-
-    // Verify the files are now uncommitted
-    env.but("--format json status -f")
-        .allow_json()
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-{
-  "uncommittedChanges": [
-    {
-      "cliId": "n",
-      "filePath": "a.txt",
-      "changeType": "added"
-    },
-    {
-      "cliId": "p",
-      "filePath": "b.txt",
-      "changeType": "added"
-    }
-  ],
-  "stacks": [
-    {
-      "cliId": "k0",
-      "assignedChanges": [],
-      "branches": [
-...
-
-"#]]);
-
-    Ok(())
-}
-
-#[test]
-fn uncommit_diff_json_keeps_mutation_result_and_diff() -> anyhow::Result<()> {
-    fn run(agent: bool) -> anyhow::Result<()> {
-        let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-
-        env.setup_metadata(&["A", "B"]);
-        commit_two_files_as_two_hunks_each(&env, "A", "a.txt", "b.txt", "first commit");
-
-        let before = status_json(&env)?;
-        let commit_cli_id = branch_commit_cli_ids(&before, "A")[0].clone();
-        let command = format!("--format json uncommit {commit_cli_id} --diff");
-        let output = if agent {
-            env.but(command)
-                .env("AI_AGENT", "codex")
-                .allow_json()
-                .output()?
-        } else {
-            env.but(command).allow_json().output()?
-        };
-        assert!(
-            output.status.success(),
-            "uncommit --diff failed: {}",
-            String::from_utf8_lossy(&output.stderr)
-        );
-
-        let json: serde_json::Value = serde_json::from_slice(&output.stdout)?;
-        let result = if agent { &json["result"] } else { &json };
-        assert_eq!(result["ok"], true);
-
-        let changes = result["diff"]["changes"]
-            .as_array()
-            .expect("diff changes should be an array");
-        assert!(
-            changes
-                .iter()
-                .any(|change| change["path"].as_str() == Some("a.txt")),
-            "diff output should include the uncommitted files"
-        );
-
-        if agent {
-            assert!(
-                json.get("status").is_some(),
-                "agent JSON output should still include status"
-            );
-        }
-
-        Ok(())
-    }
-
-    run(false)?;
-    run(true)?;
-    Ok(())
-}
-
-#[test]
-fn uncommit_command_validation() {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-
-    env.setup_metadata(&["A", "B"]);
-    commit_file_with_worktree_changes_as_two_hunks(&env, "A", "a.txt");
-
-    // Test that uncommit rejects uncommitted files
-    env.but("uncommit a.txt")
-        .assert()
-        .failure()
-        .stderr_eq(str![[r#"
-Failed to uncommit. Cannot uncommit a.txt - it is an uncommitted file or hunk. Only commits and files-in-commits can be uncommitted.
-
-"#]]);
-
-    // Test that uncommit rejects branches
-    env.but("uncommit A").assert().failure().stderr_eq(str![[r#"
-Failed to uncommit. Cannot uncommit A - it is a branch. Only commits and files-in-commits can be uncommitted.
-
-"#]]);
-}
-
-#[test]
-fn uncommit_command_with_discard_on_commit() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-
-    env.setup_metadata(&["A", "B"]);
-    commit_two_files_as_two_hunks_each(&env, "A", "a.txt", "b.txt", "first commit");
-
-    let before = status_json(&env)?;
-    let commit_cli_ids_before = branch_commit_cli_ids(&before, "A");
-    let source_cli_id = commit_cli_ids_before[0].clone();
-
-    env.but("stf")
-        .assert()
-        .success()
-        .stderr_eq(snapbox::str![])
-        .stdout_eq(snapbox::str![[r#"
-╭┄zz [uncommitted] (no changes)
-┊
-┊╭┄g0 [A]
-┊●   1 create a.txt and b.txt
-┊│     1:n A a.txt
-┊│     1:p A b.txt
-┊●   tpm add A
-┊│     tpm:t A A
-├╯
-┊
-┊╭┄h0 [B]
-┊●   lrm add B
-┊│     lrm:p A B
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
-
-    env.but(format!("uncommit {source_cli_id} --discard"))
-        .assert()
-        .success();
-
-    let after = status_json(&env)?;
-    let commit_cli_ids_after = branch_commit_cli_ids(&after, "A");
-
-    assert_eq!(
-        commit_cli_ids_after.len() + 1,
-        commit_cli_ids_before.len(),
-        "discarding a commit via uncommit should remove that commit from branch history"
-    );
-    assert!(
-        !commit_cli_ids_after.contains(&source_cli_id),
-        "source commit should no longer be present after discard"
-    );
-    assert!(
-        !uncommitted_contains_file(&after, "a.txt") && !uncommitted_contains_file(&after, "b.txt"),
-        "discarding a commit should not move its changes into uncommitted"
-    );
-
-    env.but("stf")
-        .assert()
-        .success()
-        .stderr_eq(snapbox::str![])
-        .stdout_eq(snapbox::str![[r#"
-╭┄zz [uncommitted] (no changes)
-┊
-┊╭┄g0 [A]
-┊●   tpm add A
-┊│     tpm:t A A
-├╯
-┊
-┊╭┄h0 [B]
-┊●   lrm add B
-┊│     lrm:p A B
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
-
-    Ok(())
-}
-
-#[test]
-fn uncommit_command_with_discard_on_committed_file() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-
-    env.setup_metadata(&["A", "B"]);
-    commit_two_files_as_two_hunks_each(&env, "A", "a.txt", "b.txt", "first commit");
-
-    let before = status_json(&env)?;
-    let committed_file_cli_id = committed_file_cli_id_for_file(&before, "A", "b.txt")
-        .expect("b.txt committed-file id should exist");
-
-    env.but("stf")
-        .assert()
-        .success()
-        .stderr_eq(snapbox::str![])
-        .stdout_eq(snapbox::str![[r#"
-╭┄zz [uncommitted] (no changes)
-┊
-┊╭┄g0 [A]
-┊●   1 create a.txt and b.txt
-┊│     1:n A a.txt
-┊│     1:p A b.txt
-┊●   tpm add A
-┊│     tpm:t A A
-├╯
-┊
-┊╭┄h0 [B]
-┊●   lrm add B
-┊│     lrm:p A B
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
-
-    env.but(format!("uncommit {committed_file_cli_id} -d"))
-        .assert()
-        .success();
-
-    let after = status_json(&env)?;
-    assert!(
-        !uncommitted_contains_file(&after, "b.txt"),
-        "discarded committed file changes should not end up uncommitted"
-    );
-    assert!(
-        !branch_commits_contain_file(&after, "A", "b.txt"),
-        "discarded committed file changes should no longer be in commit history"
-    );
-    assert!(
-        branch_commits_contain_file(&after, "A", "a.txt"),
-        "other committed file changes should remain in history"
-    );
-
-    env.but("stf")
-        .assert()
-        .success()
-        .stderr_eq(snapbox::str![])
-        .stdout_eq(snapbox::str![[r#"
-╭┄zz [uncommitted] (no changes)
-┊
-┊╭┄g0 [A]
-┊●   1 create a.txt and b.txt
-┊│     1:n A a.txt
-┊●   tpm add A
-┊│     tpm:t A A
-├╯
-┊
-┊╭┄h0 [B]
-┊●   lrm add B
-┊│     lrm:p A B
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
-
-    Ok(())
-}
-
-#[test]
-fn uncommit_help_mentions_discard_flag() -> anyhow::Result<()> {
-    let env = Sandbox::empty();
-
-    let output = env.but("uncommit --help").output()?;
-    assert!(output.status.success());
-    let stdout = String::from_utf8(output.stdout)?;
-
-    assert!(
-        stdout.contains("-d, --discard"),
-        "expected uncommit help to list the discard flag"
-    );
-    assert!(
-        stdout.contains("Discard the selected committed changes"),
-        "expected uncommit help to describe discard behavior"
-    );
-
-    Ok(())
-}
-
-#[test]
-fn agent_uncommit_discard_multiple_sources_writes_single_json_with_status() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-
-    env.setup_metadata(&["A", "B"]);
-    commit_two_files_as_two_hunks_each(&env, "A", "a.txt", "b.txt", "first commit");
-    commit_two_files_as_two_hunks_each(&env, "A", "a.txt", "b.txt", "second commit");
-
-    let before = status_json(&env)?;
-    let commit_cli_ids_before = branch_commit_cli_ids(&before, "A");
-    let source_cli_ids = format!("{},{}", commit_cli_ids_before[0], commit_cli_ids_before[1]);
-
-    let output = env
-        .but(format!("--format json uncommit {source_cli_ids} --discard"))
-        .env("AI_AGENT", "codex")
-        .allow_json()
-        .output()?;
-    assert!(output.status.success());
-
-    let parsed: serde_json::Value = serde_json::from_slice(&output.stdout)?;
-    assert_eq!(parsed["result"]["ok"], serde_json::json!(true));
-    assert!(
-        parsed.get("status").is_some(),
-        "agent JSON wrapper should include status"
-    );
-
-    let after = status_json(&env)?;
-    let commit_cli_ids_after = branch_commit_cli_ids(&after, "A");
-
-    assert_eq!(
-        commit_cli_ids_after.len() + 2,
-        commit_cli_ids_before.len(),
-        "discarding two commit sources should remove both from branch history"
-    );
-    assert!(
-        !uncommitted_contains_file(&after, "a.txt") && !uncommitted_contains_file(&after, "b.txt"),
-        "discarded commits should not move their changes into uncommitted"
-    );
-
-    Ok(())
-}
-
-#[test]
-fn stage_command() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-
-    env.setup_metadata(&["A", "B"]);
-    commit_file_with_worktree_changes_as_two_hunks(&env, "A", "a.txt");
-
-    // Test stage command
-    env.but("stage a.txt A").assert().success();
-
-    // Verify the file is assigned to A
-    env.but("--format json status -f")
-        .allow_json()
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-{
-  "uncommittedChanges": [],
-  "stacks": [
-    {
-      "cliId": "j0",
-      "assignedChanges": [
-        {
-          "cliId": "k",
-          "filePath": "a.txt",
-          "changeType": "modified"
-        }
-      ],
-...
-
-"#]]);
-
-    Ok(())
-}
-
-#[test]
-fn stage_command_path_prefix() {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-
-    env.setup_metadata(&["A", "B"]);
-    env.file("path/a.txt", "text\n");
-    env.but("stage path/ A")
-        .assert()
-        .success()
-        .stderr_eq(snapbox::str![])
-        .stdout_eq(snapbox::str![[r#"
-Staged hunk(s) → [A].
-
-"#]]);
-}
-
-#[test]
-fn stage_command_missing_source_hints_to_refresh_cli_ids() {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-
-    env.setup_metadata(&["A", "B"]);
-    env.but("stage missing-file A")
-        .assert()
-        .failure()
-        .stderr_eq(str![[r#"
-Error: Bad input 'missing-file' for '<FILE_OR_HUNK>'
-
-Source 'missing-file' not found. If you just performed a Git operation (squash, rebase, etc.), try running 'but status' to refresh the current state.
-
-Hint: Run `but status --format json -f` to refresh CLI IDs, then retry with a file or hunk cliId from the output
-
-"#]]);
-}
-
-#[test]
-fn stage_command_missing_branch_hints_to_refresh_cli_ids() {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-
-    env.setup_metadata(&["A", "B"]);
-    env.file("a.txt", "text\n");
-
-    env.but("stage a.txt missing-branch")
-        .assert()
-        .failure()
-        .stderr_eq(str![[r#"
-Error: Bad input 'missing-branch' for '<BRANCH>'
-
-Branch 'missing-branch' not found. If you just performed a Git operation (squash, rebase, etc.), try running 'but status' to refresh the current state.
-
-Hint: Use a branch name or branch cliId from `but status --format json -f`
-
-"#]]);
-}
-
-#[test]
-fn stage_command_non_branch_target_hints_to_use_branch() {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-
-    env.setup_metadata(&["A", "B"]);
-    env.file("a.txt", "text\n");
-
-    env.but("stage a.txt zz")
-        .assert()
-        .failure()
-        .stderr_eq(str![[r#"
-Error: Bad input 'zz' for '<BRANCH>'
-
-Cannot stage to zz - it is the uncommitted area. Target must be a branch.
-
-Hint: Use a branch name or branch cliId from `but status --format json -f`
-
-"#]]);
-}
-
-#[test]
-fn unstage_command() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-
-    env.setup_metadata(&["A", "B"]);
-    commit_file_with_worktree_changes_as_two_hunks(&env, "A", "a.txt");
-
-    // First stage the file to A
-    env.but("stage a.txt A").assert().success();
-
-    // Verify it's assigned
-    env.but("--format json status -f")
-        .allow_json()
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-{
-  "uncommittedChanges": [],
-  "stacks": [
-    {
-      "cliId": "j0",
-      "assignedChanges": [
-        {
-          "cliId": "k",
-          "filePath": "a.txt",
-          "changeType": "modified"
-        }
-      ],
-...
-
-"#]]);
-
-    // Now unstage it
-    env.but("unstage A@{stack}:a.txt").assert().success();
-
-    // Verify it's now uncommitted
-    env.but("--format json status -f")
-        .allow_json()
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-{
-  "uncommittedChanges": [
-    {
-      "cliId": "n",
-      "filePath": "a.txt",
-      "changeType": "modified"
-    }
-  ],
-  "stacks": [
-    {
-      "cliId": "j0",
-      "assignedChanges": [],
-...
-
-"#]]);
-
-    Ok(())
-}
-
-#[test]
-fn unstage_command_path_prefix() {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-
-    env.setup_metadata(&["A", "B"]);
-    env.file("path/a.txt", "text\n");
-
-    // First stage the file to A
-    env.but("stage path/a.txt A").assert().success();
-
-    // Now unstage it, giving a path prefix
-    env.but("unstage A@{stack}:path/ A")
-        .assert()
-        .success()
-        .stderr_eq(snapbox::str![])
-        .stdout_eq(snapbox::str![[r#"
-Unstaged hunk(s)
-
-"#]]);
-}
-
-#[test]
-fn unstage_command_with_branch() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-
-    env.setup_metadata(&["A", "B"]);
-    commit_file_with_worktree_changes_as_two_hunks(&env, "A", "a.txt");
-
-    // Stage the file to A
-    env.but("stage a.txt A").assert().success();
-
-    // Unstage with branch parameter
-    env.but("unstage A@{stack}:a.txt A").assert().success();
-
-    // Verify it's uncommitted
-    env.but("--format json status -f")
-        .allow_json()
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-{
-  "uncommittedChanges": [
-    {
-      "cliId": "n",
-      "filePath": "a.txt",
-      "changeType": "modified"
-    }
-  ],
-...
-
-"#]]);
-
-    Ok(())
-}
-
-#[test]
-fn unstage_command_validation() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-
-    env.setup_metadata(&["A", "B"]);
-    commit_two_files_as_two_hunks_each(&env, "A", "a.txt", "b.txt", "first commit");
-
-    // Get the commit CLI ID from status
-    let status_output = env.but("--format json status").allow_json().output()?;
-    let status: serde_json::Value = serde_json::from_slice(&status_output.stdout)?;
-    let commit_cli_id = status["stacks"][0]["branches"][0]["commits"][0]["cliId"]
-        .as_str()
-        .unwrap();
-
-    // Test that unstage rejects commits
-    env.but(format!("unstage {commit_cli_id}"))
-        .assert()
-        .failure()
-        .stderr_eq(str![[r#"
-Failed to unstage. '1' is a commit but must be an uncommitted file or hunk
-
-"#]]);
-
-    // Test that unstage rejects non-branch as branch parameter. Refresh the ID after adding a
-    // commit, as duplicate change IDs can gain a new disambiguator.
-    commit_file_with_worktree_changes_as_two_hunks(&env, "A", "c.txt");
-    let commit_cli_id = branch_commit_cli_ids(&status_json(&env)?, "A")[0].clone();
-    env.but(format!("unstage c.txt {commit_cli_id}"))
-        .assert()
-        .failure()
-        .stderr_eq(str![[r#"
-Failed to unstage. Cannot unstage from 2 - it is a commit. Target must be a branch.
-
-"#]]);
-
-    Ok(())
-}
-
-// Full rub matrix CLI smoke coverage.
-
-#[test]
-fn rub_matrix_uncommitted_hunk_to_commit_smoke() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-    env.setup_metadata(&["A", "B"]);
-
-    env.file("uncommitted-to-commit.txt", "content\n");
-
-    let before = status_json(&env)?;
-    let target_cli_id = branch_commit_cli_ids(&before, "A")[0].clone();
-
-    env.but(format!("rub uncommitted-to-commit.txt {target_cli_id}"))
-        .assert()
-        .success()
-        .stdout_eq(str![[r#"
-Amended [..] → [..]
-
-"#]])
-        .stderr_eq(str![""]);
-
-    let after = status_json(&env)?;
-    assert!(
-        !uncommitted_contains_file(&after, "uncommitted-to-commit.txt"),
-        "file should no longer be uncommitted"
-    );
-    assert!(
-        branch_commits_contain_file(&after, "A", "uncommitted-to-commit.txt"),
-        "file should appear in commits on branch A"
-    );
-
-    Ok(())
 }
 
 #[test]
@@ -1297,199 +107,6 @@ Staged all unstaged changes to [A].
 }
 
 #[test]
-fn rub_matrix_uncommitted_area_to_commit_smoke() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-    env.setup_metadata(&["A", "B"]);
-
-    env.file("zz-to-commit.txt", "content\n");
-
-    let before = status_json(&env)?;
-    let target_cli_id = branch_commit_cli_ids(&before, "A")[0].clone();
-
-    env.but(format!("rub zz {target_cli_id}"))
-        .assert()
-        .success()
-        .stdout_eq(str![[r#"
-Amended uncommitted files → [..]
-
-"#]])
-        .stderr_eq(str![""]);
-
-    let after = status_json(&env)?;
-    assert!(
-        !uncommitted_contains_file(&after, "zz-to-commit.txt"),
-        "file should no longer be uncommitted"
-    );
-    assert!(
-        branch_commits_contain_file(&after, "A", "zz-to-commit.txt"),
-        "file should appear in commits on branch A"
-    );
-
-    Ok(())
-}
-
-#[test]
-fn rub_matrix_uncommitted_to_commit_consumes_renames() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-    env.setup_metadata(&["A", "B"]);
-
-    let original = (1..=120)
-        .map(|line| line.to_string())
-        .collect::<Vec<_>>()
-        .join("\n")
-        + "\n";
-    env.file("rename-source.txt", &original);
-    env.but("commit A -m 'seed rename source'")
-        .assert()
-        .success();
-
-    std::fs::rename(
-        env.projects_root().join("rename-source.txt"),
-        env.projects_root().join("rename-target.txt"),
-    )?;
-    env.file(
-        "rename-target.txt",
-        original.replace("40\n41\n42\n", "40\nchanged\n42\n"),
-    );
-
-    let before = status_json(&env)?;
-    let target_cli_id = branch_commit_cli_ids(&before, "A")[0].clone();
-
-    env.but(format!("rub zz {target_cli_id}"))
-        .assert()
-        .success()
-        .stdout_eq(str![[r#"
-Amended uncommitted files → [..]
-
-"#]])
-        .stderr_eq(str![""]);
-
-    let after = status_json(&env)?;
-    assert!(
-        !uncommitted_contains_file(&after, "rename-target.txt"),
-        "renamed file should no longer be uncommitted"
-    );
-    assert_eq!(
-        env.invoke_git("status --porcelain"),
-        "",
-        "expected all zz changes to be committed"
-    );
-
-    Ok(())
-}
-
-#[test]
-fn rub_matrix_uncommitted_file_to_commit_consumes_renames() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-    env.setup_metadata(&["A", "B"]);
-
-    let original = (1..=120)
-        .map(|line| line.to_string())
-        .collect::<Vec<_>>()
-        .join("\n")
-        + "\n";
-    env.file("rename-source-single.txt", &original);
-    env.but("commit A -m 'seed rename source single'")
-        .assert()
-        .success();
-
-    std::fs::rename(
-        env.projects_root().join("rename-source-single.txt"),
-        env.projects_root().join("rename-target-single.txt"),
-    )?;
-    env.file(
-        "rename-target-single.txt",
-        original.replace("70\n71\n72\n", "70\nchanged\n72\n"),
-    );
-
-    let before = status_json(&env)?;
-    let source_file_cli_id = uncommitted_cli_id_for_file(&before, "rename-target-single.txt")
-        .expect("renamed uncommitted file should be present in status");
-    let target_cli_id = branch_commit_cli_ids(&before, "A")[0].clone();
-
-    env.but(format!("rub {source_file_cli_id} {target_cli_id}"))
-        .assert()
-        .success()
-        .stdout_eq(str![[r#"
-Amended the only hunk in rename-target-single.txt in the uncommitted area → [..]
-
-"#]])
-        .stderr_eq(str![""]);
-
-    let after = status_json(&env)?;
-    assert!(
-        !uncommitted_contains_file(&after, "rename-target-single.txt"),
-        "renamed file should no longer be uncommitted"
-    );
-
-    let remaining = env.invoke_git("status --porcelain");
-    assert_eq!(
-        remaining, "",
-        "expected selected renamed file to be committed; remaining status:\n{remaining}"
-    );
-
-    Ok(())
-}
-
-#[test]
-fn rub_uncommitted_deleted_file_to_commit_keeps_unrelated_deleted_file() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-    env.setup_metadata(&["A", "B"]);
-
-    env.file("a.txt", "a\n");
-    env.file("b.txt", "b\n");
-    env.file("c.txt", "c\n");
-    env.but("commit A -m 'Add a.txt, b.txt, and c.txt'")
-        .assert()
-        .success();
-
-    std::fs::remove_file(env.projects_root().join("a.txt"))?;
-    std::fs::remove_file(env.projects_root().join("b.txt"))?;
-
-    let before = status_json(&env)?;
-    let source_file_cli_id = uncommitted_cli_id_for_file(&before, "a.txt")
-        .expect("a.txt deletion should be present in the uncommitted area");
-    let target_cli_id = branch_commit_cli_ids(&before, "A")[0].clone();
-    assert!(
-        uncommitted_contains_file(&before, "b.txt"),
-        "b.txt deletion should start in the uncommitted area"
-    );
-
-    env.but(format!("rub {source_file_cli_id} {target_cli_id}"))
-        .assert()
-        .success()
-        .stdout_eq(str![[r#"
-Amended the only hunk in a.txt in the uncommitted area → [..]
-
-"#]])
-        .stderr_eq(str![""]);
-
-    let after = status_json(&env)?;
-    assert!(
-        !uncommitted_contains_file(&after, "a.txt"),
-        "selected a.txt deletion should be amended into the target commit"
-    );
-    assert!(
-        uncommitted_contains_file(&after, "b.txt"),
-        "unrelated b.txt deletion should remain uncommitted"
-    );
-    assert!(
-        !env.projects_root().join("a.txt").exists(),
-        "selected a.txt deletion should stay applied to the worktree"
-    );
-    assert!(
-        !env.projects_root().join("b.txt").exists(),
-        "unrelated b.txt deletion should stay applied to the worktree"
-    );
-    assert!(
-        env.projects_root().join("c.txt").exists(),
-        "untouched c.txt should stay in the worktree"
-    );
-
-    Ok(())
-}
-
-#[test]
 fn rub_matrix_uncommitted_area_to_stack_smoke() -> anyhow::Result<()> {
     let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
     env.setup_metadata(&["A", "B"]);
@@ -1513,200 +130,6 @@ Staged all unstaged changes to [A].
     assert!(
         stack_assigned_contains_file(&after, "A", "zz-to-stack.txt"),
         "file should be assigned to A stack"
-    );
-
-    Ok(())
-}
-
-#[test]
-fn rub_matrix_commit_to_uncommitted_smoke() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-    env.setup_metadata(&["A", "B"]);
-
-    commit_two_files_as_two_hunks_each(&env, "A", "a.txt", "b.txt", "first commit");
-
-    let before = status_json(&env)?;
-    let commit_cli_ids_before = branch_commit_cli_ids(&before, "A");
-    let source_cli_id = commit_cli_ids_before[0].clone();
-
-    env.but(format!("rub {source_cli_id} zz"))
-        .assert()
-        .success()
-        .stdout_eq(str![[r#"
-Uncommitted [..]
-
-"#]])
-        .stderr_eq(str![""]);
-
-    let after = status_json(&env)?;
-    let commit_cli_ids_after = branch_commit_cli_ids(&after, "A");
-
-    assert_eq!(
-        commit_cli_ids_after.len() + 1,
-        commit_cli_ids_before.len(),
-        "uncommitting a commit should remove that commit from branch history"
-    );
-    assert!(
-        !commit_cli_ids_after.contains(&source_cli_id),
-        "source commit should no longer be present after uncommit"
-    );
-
-    assert!(
-        uncommitted_contains_file(&after, "a.txt") && uncommitted_contains_file(&after, "b.txt"),
-        "uncommitting a commit should move its changes into uncommitted"
-    );
-
-    Ok(())
-}
-
-#[test]
-fn rub_matrix_commit_to_commit_smoke() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-    env.setup_metadata(&["A", "B"]);
-
-    commit_two_files_as_two_hunks_each(&env, "A", "a.txt", "b.txt", "first commit");
-    commit_two_files_as_two_hunks_each(&env, "A", "a.txt", "b.txt", "second commit");
-
-    let before = status_json(&env)?;
-    let commit_cli_ids_before = branch_commit_cli_ids(&before, "A");
-    let source_cli_id = commit_cli_ids_before[0].clone();
-    let target_cli_id = commit_cli_ids_before[1].clone();
-
-    env.but(format!("rub {source_cli_id} {target_cli_id}"))
-        .assert()
-        .success()
-        .stdout_eq(str![[r#"
-Squashed [..] → [..]
-
-"#]])
-        .stderr_eq(str![""]);
-
-    let after = status_json(&env)?;
-    let commit_cli_ids_after = branch_commit_cli_ids(&after, "A");
-    assert_eq!(
-        commit_cli_ids_after.len() + 1,
-        commit_cli_ids_before.len(),
-        "squashing should reduce commit count by one"
-    );
-
-    Ok(())
-}
-
-#[test]
-fn rub_commit_without_message_to_commit() {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("one-stack");
-    env.setup_metadata(&["A"]);
-
-    env.file("one.txt", "one.txt contents");
-    env.but("commit -m 'add one.txt'").assert().success();
-
-    env.but("status --no-hint")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄zz [uncommitted] (no changes)
-┊
-┊╭┄g0 [A]
-┊●   1 add one.txt
-┊●   tpm add A
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-"#]]);
-
-    env.but("commit empty --after 1").assert().success();
-
-    env.but("status --no-hint")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄zz [uncommitted] (no changes)
-┊
-┊╭┄g0 [A]
-┊●   1#0 (no commit message) (no changes)
-┊●   1#1 add one.txt
-┊●   tpm add A
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-"#]]);
-
-    env.but("rub 1#0 1#1").assert().success();
-
-    env.but("status --no-hint")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄zz [uncommitted] (no changes)
-┊
-┊╭┄g0 [A]
-┊●   1 add one.txt
-┊●   tpm add A
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-"#]]);
-}
-
-#[test]
-fn rub_commit_to_commit_without_message() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("one-stack");
-    env.setup_metadata(&["A"]);
-
-    env.file("one.txt", "one.txt contents");
-    env.but("commit -m 'add one.txt'").assert().success();
-    env.but("commit empty --after 1").assert().success();
-
-    env.but("rub 1#1 1#0").assert().success();
-
-    let status = status_json(&env)?;
-    let branch = status["stacks"]
-        .as_array()
-        .unwrap()
-        .iter()
-        .flat_map(|stack| stack["branches"].as_array().unwrap().iter())
-        .find(|branch| branch["name"].as_str().unwrap() == "A")
-        .unwrap();
-    let commit_messages = branch["commits"]
-        .as_array()
-        .unwrap()
-        .iter()
-        .map(|commit| commit["message"].as_str().unwrap().trim_end_matches('\n'))
-        .collect::<Vec<_>>();
-
-    assert_eq!(commit_messages, vec!["add one.txt", "add A"]);
-
-    Ok(())
-}
-
-#[test]
-fn rub_matrix_commit_to_branch_smoke() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-    env.setup_metadata(&["A", "B"]);
-
-    commit_two_files_as_two_hunks_each(&env, "A", "a.txt", "b.txt", "first commit");
-
-    let before = status_json(&env)?;
-    let source_cli_id = branch_commit_cli_ids(&before, "A")[0].clone();
-    let branch_b_count_before = branch_commit_cli_ids(&before, "B").len();
-
-    env.but(format!("rub {source_cli_id} B"))
-        .assert()
-        .success()
-        .stdout_eq(str![[r#"
-Moved [..] → [B]
-
-"#]])
-        .stderr_eq(str![""]);
-
-    let after = status_json(&env)?;
-    let branch_b_count_after = branch_commit_cli_ids(&after, "B").len();
-    assert!(
-        branch_b_count_after > branch_b_count_before,
-        "moving a commit to B should increase commit count on B"
     );
 
     Ok(())
@@ -2049,38 +472,6 @@ Uncommitted changes
 }
 
 #[test]
-fn rub_matrix_committed_file_to_commit_smoke() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("one-stack");
-    env.setup_metadata(&["A"]);
-
-    commit_two_files_as_two_hunks_each(&env, "A", "source-a.txt", "source-b.txt", "source commit");
-    commit_two_files_as_two_hunks_each(&env, "A", "target-a.txt", "target-b.txt", "target commit");
-
-    let before = status_json(&env)?;
-    let source_cli_id = branch_commit_cli_id_for_file(&before, "A", "source-a.txt")
-        .expect("source commit with file");
-    let target_cli_id = branch_commit_cli_id_for_file(&before, "A", "target-a.txt")
-        .expect("target commit with file");
-
-    env.but(format!("rub {source_cli_id}:source-a.txt {target_cli_id}"))
-        .assert()
-        .success()
-        .stdout_eq(str![[r#"
-Moved files between commits!
-
-"#]])
-        .stderr_eq(str![""]);
-
-    let after = status_json(&env)?;
-    assert!(
-        branch_commits_contain_file(&after, "A", "source-a.txt"),
-        "file should still be present in branch A history after moving between commits"
-    );
-
-    Ok(())
-}
-
-#[test]
 fn rub_matrix_invalid_pairs_smoke() -> anyhow::Result<()> {
     let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
     env.setup_metadata(&["A", "B"]);
@@ -2122,249 +513,4 @@ Rubbed the wrong way. Operation doesn't make sense.[..]
 "#]]);
 
     Ok(())
-}
-
-#[test]
-fn agent_json_wraps_mutation_and_status() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-
-    env.setup_metadata(&["A", "B"]);
-    env.file("a.txt", "arbitrary text\n");
-
-    let output = env
-        .but("--format json stage a.txt A")
-        .env("AI_AGENT", "codex")
-        .allow_json()
-        .output()?;
-    assert!(output.status.success());
-
-    let json: serde_json::Value = serde_json::from_slice(&output.stdout)?;
-
-    // The combined output must have both "result" and "status" fields
-    assert!(
-        json.get("result").is_some(),
-        "expected 'result' field in combined JSON output"
-    );
-    assert!(
-        json.get("status").is_some(),
-        "expected 'status' field in combined JSON output"
-    );
-
-    // The result should contain the mutation output (stage produces {"ok": true})
-    assert_eq!(
-        json["result"]["ok"], true,
-        "mutation result should indicate success"
-    );
-
-    // The status should have standard status fields
-    assert!(
-        json["status"].get("stacks").is_some(),
-        "status should contain 'stacks'"
-    );
-    assert!(
-        json["status"].get("uncommittedChanges").is_some(),
-        "status should contain 'uncommittedChanges'"
-    );
-
-    Ok(())
-}
-
-#[test]
-fn agent_invocation_enables_status_after_for_mutations() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-
-    env.setup_metadata(&["A", "B"]);
-    env.file("agent.txt", "content\n");
-
-    let output = env
-        .but("--format json stage agent.txt A")
-        .env("AI_AGENT", "codex")
-        .allow_json()
-        .output()?;
-    assert!(output.status.success());
-
-    let json: serde_json::Value = serde_json::from_slice(&output.stdout)?;
-    assert!(
-        json.get("result").is_some(),
-        "agent mutation output should include the command result"
-    );
-    assert!(
-        json.get("status").is_some(),
-        "agent mutation output should include workspace status"
-    );
-
-    Ok(())
-}
-
-#[test]
-fn agent_json_success_has_no_status_error_field() -> anyhow::Result<()> {
-    // Verifies that on a successful agent mutation, the combined JSON output
-    // contains {result, status} but NOT status_error.
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-
-    env.setup_metadata(&["A", "B"]);
-    env.file("b.txt", "content\n");
-
-    let output = env
-        .but("--format json stage b.txt A")
-        .env("AI_AGENT", "codex")
-        .allow_json()
-        .output()?;
-    assert!(output.status.success());
-
-    let json: serde_json::Value = serde_json::from_slice(&output.stdout)?;
-
-    // On success, status_error should NOT be present
-    assert!(
-        json.get("status_error").is_none(),
-        "status_error should not be present on success"
-    );
-
-    Ok(())
-}
-
-#[test]
-fn rubbing_modified_and_renamed_file() {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("zero-stacks");
-    env.setup_metadata(&[]);
-
-    env.file("file", "content");
-    env.file("file-2", "content-2");
-
-    env.but("commit -m 'add files'").assert().success();
-
-    env.but("status -f")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄zz [uncommitted] (no changes)
-┊
-┊╭┄br [a-branch-1]
-┊●   1 add files
-┊│     1:q A file
-┊│     1:k A file-2
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
-
-    env.file("file-2", "new content");
-    env.rename_file("file-2", "file");
-
-    env.but("status -f")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄zz [uncommitted]
-┊   q M file
-┊   k D file-2
-┊
-┊╭┄br [a-branch-1]
-┊●   1 add files
-┊│     1:q A file
-┊│     1:k A file-2
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but diff` to see uncommitted changes and `but commit <branch> -m "message" --changes <id>` to commit them
-
-"#]]);
-
-    env.but("rub zz 1").assert().success();
-
-    env.but("status -f")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄zz [uncommitted] (no changes)
-┊
-┊╭┄br [a-branch-1]
-┊●   1 add files
-┊│     1:q A file
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
-}
-
-#[test]
-fn committing_modified_and_renamed_file() {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("zero-stacks");
-    env.setup_metadata(&[]);
-
-    env.file("file", "content");
-    env.file("file-2", "content-2");
-
-    env.but("commit -m 'add files'").assert().success();
-
-    env.but("status -f")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄zz [uncommitted] (no changes)
-┊
-┊╭┄br [a-branch-1]
-┊●   1 add files
-┊│     1:q A file
-┊│     1:k A file-2
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
-
-    env.file("file-2", "new content");
-    env.rename_file("file-2", "file");
-
-    env.but("status -f")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄zz [uncommitted]
-┊   q M file
-┊   k D file-2
-┊
-┊╭┄br [a-branch-1]
-┊●   1 add files
-┊│     1:q A file
-┊│     1:k A file-2
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but diff` to see uncommitted changes and `but commit <branch> -m "message" --changes <id>` to commit them
-
-"#]]);
-
-    env.but("commit -m 'change file'").assert().success();
-
-    env.but("status -f")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄zz [uncommitted] (no changes)
-┊
-┊╭┄br [a-branch-1]
-┊●   1#0 change file
-┊│     1#0:q M file
-┊│     1#0:k D file-2
-┊●   1#1 add files
-┊│     1#1:q A file
-┊│     1#1:k A file-2
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
 }
